% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10am.hfrs-package.R
\docType{package}
\name{icd10am.hfrs-package}
\alias{icd10am.hfrs}
\alias{icd10am.hfrs-package}
\title{icd10am.hfrs: Hospital Frailty Risk Score for ICD-10-AM Coded Data}
\description{
Calculates the Hospital Frailty Risk Score (HFRS) for Australian admitted patient care data coded using ICD-10-AM, based on the method of Gilbert et al. (2018).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/healthpolicy/icd10am.hfrs}
  \item Report bugs at \url{https://github.com/healthpolicy/icd10am.hfrs/issues}
}

}
\author{
\strong{Maintainer}: Jim Pearse \email{ejpearse@gmail.com}

}
\keyword{internal}
